fix: provide as extractor
