+++ packages/extractor-arbitrary-variants/package.json
@@ -1,22 +1,22 @@
 {
-  "name": "@unocss/preset-arbitrary",
+  "name": "@unocss/extractor-arbitrary-variants",
   "version": "0.50.6",
-  "description": "Support arbitrary variants for utilities",
+  "description": "Extractor arbitrary variants for utilities",
   "author": "Anthony Fu <anthonyfu117@hotmail.com>",
   "license": "MIT",
   "funding": "https://github.com/sponsors/antfu",
-  "homepage": "https://github.com/unocss/unocss/tree/main/packages/preset-arbitrary#readme",
+  "homepage": "https://github.com/unocss/unocss/tree/main/packages/extractor-arbitrary-variants#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/unocss/unocss.git",
-    "directory": "packages/preset-arbitrary"
+    "directory": "packages/extractor-arbitrary-variants"
   },
   "bugs": {
     "url": "https://github.com/unocss/unocss/issues"
   },
   "keywords": [
     "unocss",
-    "unocss-preset"
+    "unocss-extractor"
   ],
   "sideEffects": false,
   "exports": {
