+++ packages/shared-common/src/index.ts
@@ -1,7 +1,7 @@
 import type { ExtractorContext, UnoGenerator } from '@unocss/core'
 import { escapeRegExp, isAttributifySelector, makeRegexClassGroup, splitWithVariantGroupRE } from '@unocss/core'
 import MagicString from 'magic-string'
-import { arbitraryPropertyRE, quotedArbitraryValuesRE } from '../../preset-arbitrary/src/split-arbitrary'
+import { arbitraryPropertyRE, quotedArbitraryValuesRE } from '../../extractor-arbitrary-variants/src'
 
 // https://github.com/dsblv/string-replace-async/blob/main/index.js
 export function replaceAsync(string: string, searchValue: RegExp, replacer: (...args: string[]) => Promise<string>) {
