--- packages/unocss/src/preset-arbitrary.ts
@@ -1,2 +1,0 @@
-export * from '@unocss/preset-arbitrary'
-export { default } from '@unocss/preset-arbitrary'
