+++ test/preset-attributify.test.ts
@@ -1,6 +1,5 @@
 import { createGenerator, toEscapedSelector as e } from '@unocss/core'
 import presetUno from '@unocss/preset-uno'
-import presetArbitrary from '@unocss/preset-arbitrary'
 import presetAttributify, { autocompleteExtractorAttributify, variantAttributify } from '@unocss/preset-attributify'
 import { describe, expect, test } from 'vitest'
 
@@ -295,7 +294,6 @@ describe('attributify', () => {
       presets: [
         presetAttributify({ strict: true }),
         presetUno({ attributifyPseudo: true }),
-        presetArbitrary(),
       ],
     })
